Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11001 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 281d9586c2df84cbbe52f7bd02c53114ab9bb123

+++ sensor/objecttracking/src/de/uniol/inf/is/odysseus/objecttracking/physicaloperator/access/AtomicDataInputStreamAccessMVPO.java
@@ -182,7 +182,7 @@ public class AtomicDataInputStreamAccessMVPO<M extends IProbability> extends Abs
 			this.isDone = true;
 			this.end = System.nanoTime();
 			long duration = (end - start);
-			long elems = (long)this.elemCount * 1000L * 1000L * 1000L;
+			long elems = this.elemCount * 1000L * 1000L * 1000L;
 			System.out.println("elements/sec: " + ((double)(elems)/ (double)(duration)));
 			
 			propagateDone();
